+++ RR.AI-Chat/RR.AI-Chat.Repository/Migrations/AIChatDbContextModelSnapshot.cs
@@ -225,15 +225,15 @@ namespace RR.AI_Chat.Repository.Migrations
                     b.Property<DateTime>("DateModified")
                         .HasColumnType("datetime2");
 
-                    b.Property<int>("InputTokens")
-                        .HasColumnType("int");
+                    b.Property<long>("InputTokens")
+                        .HasColumnType("bigint");
 
                     b.Property<string>("Name")
                         .HasMaxLength(100)
                         .HasColumnType("nvarchar(100)");
 
-                    b.Property<int>("OutputTokens")
-                        .HasColumnType("int");
+                    b.Property<long>("OutputTokens")
+                        .HasColumnType("bigint");
 
                     b.HasKey("Id");
